long pointers (for x64) and draw polymorphism

+++ LogParser/DriverCall/Functions/ClearDepthStencilView.cs
@@ -6,9 +6,9 @@
         {
         }
 
-        public uint pDepthStencilView { get; set; }
+        public ulong pDepthStencilView { get; set; }
         public uint ClearFlags { get; set; }
         public float Depth { get; set; }
-        public uint Stencil { get; set; }
+        public ulong Stencil { get; set; }
     }
 }
