+++ LogParser/Slot/ResourceView.cs
@@ -4,7 +4,7 @@ namespace Migoto.Log.Parser.Slot
 {
     public class ResourceView : Resource
     {
-        public uint View { get; set; }
+        public ulong View { get; set; }
 
         public ResourceView(Base owner) : base(owner)
         {
